+++ src/ts/WebView.ts
@@ -292,11 +292,17 @@ export class WebView {
 
     private onAdUnitStartProcessed(): void {
         if(this._currentAdUnit) {
+            // A/B test for 1 second refresh after start
+            let magicConstant: number = this._startRefreshMagicConstant;
+            if(this._currentAdUnit.getCampaign().getAbGroup() === 6) {
+                magicConstant = 1000;
+            }
+
             setTimeout(() => {
                 if(!this._mustReinitialize && this._currentAdUnit && this._currentAdUnit.isCached()) {
                     this._campaignRefreshManager.refresh();
                 }
-            }, this._startRefreshMagicConstant);
+            }, magicConstant);
         }
     }